requirements.txt
av
display
easyocr
googletrans==4.0.0-rc1
gttS
javascript
langdetect
numpy
opencv-python
pandas
pillow
playsound
pytesseract
pyttsx3
requests
sacremoses
sentencepiece
streamlit
streamlit-webrtc
tensorflow
torch
torchvision
transformers
ultralytics